+++ src/errors/chrono.rs
@@ -1,46 +1,90 @@
 use chrono::{RoundingError, ParseMonthError, ParseWeekdayError, OutOfRangeError, ParseError};
 use chrono::format::ParseErrorKind;
-use crate::{SystemErrorCodes};
+use crate::{SystemErrorCodes, TheErrorType};
 
-impl From<RoundingError> for SystemErrorCodes {
+impl From<RoundingError> for TheErrorType {
     fn from(value: RoundingError) -> Self {
+        let error_type;
+        let error_content = value.to_string();
         match value {
-            RoundingError::DurationExceedsTimestamp => {SystemErrorCodes::BadDateFormat }
-            RoundingError::DurationExceedsLimit => {SystemErrorCodes::BadDateFormat }
-            RoundingError::TimestampExceedsLimit => {SystemErrorCodes::BadDateFormat }
+            RoundingError::DurationExceedsTimestamp => {
+                error_type = SystemErrorCodes::BadDateFormat
+            }
+            RoundingError::DurationExceedsLimit => {
+                error_type = SystemErrorCodes::BadDateFormat
+            }
+            RoundingError::TimestampExceedsLimit => {
+                error_type = SystemErrorCodes::BadDateFormat
+            }
+        }
+        Self {
+            error_type,
+            error_content
         }
     }
 }
 
-impl From<ParseMonthError> for SystemErrorCodes {
-    fn from(_: ParseMonthError) -> Self {
-        SystemErrorCodes::BadDateFormat
+impl From<ParseMonthError> for TheErrorType {
+    fn from(value: ParseMonthError) -> Self {
+        Self{
+            error_type: SystemErrorCodes::BadDateFormat,
+            error_content: value.to_string()
+        }
     }
 }
 
-impl From<ParseWeekdayError> for SystemErrorCodes {
-    fn from(_: ParseWeekdayError) -> Self {
-        SystemErrorCodes::BadDateFormat
+impl From<ParseWeekdayError> for TheErrorType {
+    fn from(value: ParseWeekdayError) -> Self {
+        Self {
+            error_type: SystemErrorCodes::BadDateFormat,
+            error_content: value.to_string()
+        }
     }
 }
 
-impl From<OutOfRangeError> for SystemErrorCodes {
-    fn from(_: OutOfRangeError) -> Self {
-        SystemErrorCodes::OutOfRangeDateValue
+impl From<OutOfRangeError> for TheErrorType {
+    fn from(value: OutOfRangeError) -> Self {
+        Self {
+            error_type: SystemErrorCodes::OutOfRangeDateValue,
+            error_content: value.to_string()
+        }
     }
 }
 
-impl From<ParseError> for SystemErrorCodes {
+impl From<ParseError> for TheErrorType {
     fn from(value: ParseError) -> Self {
+        let error_type;
+        let error_content = value.to_string();
         match value.kind() {
-            ParseErrorKind::OutOfRange => {SystemErrorCodes::OutOfRangeDateValue}
-            ParseErrorKind::Impossible => {SystemErrorCodes::InvalidDateValue}
-            ParseErrorKind::NotEnough => {SystemErrorCodes::InvalidDateValue}
-            ParseErrorKind::Invalid => {SystemErrorCodes::InvalidDateValue}
-            ParseErrorKind::TooShort => {SystemErrorCodes::InvalidDateValue}
-            ParseErrorKind::TooLong => {SystemErrorCodes::InvalidDateValue}
-            ParseErrorKind::BadFormat => {SystemErrorCodes::BadDateFormat}
-            ParseErrorKind::__Nonexhaustive => {SystemErrorCodes::BadDateFormat}
+            ParseErrorKind::OutOfRange => {
+                error_type = SystemErrorCodes::OutOfRangeDateValue
+            }
+            ParseErrorKind::Impossible => {
+                error_type = SystemErrorCodes::InvalidDateValue
+            }
+            ParseErrorKind::NotEnough => {
+                error_type = SystemErrorCodes::InvalidDateValue
+            }
+            ParseErrorKind::Invalid => {
+                error_type = SystemErrorCodes::InvalidDateValue
+            }
+            ParseErrorKind::TooShort => {
+                error_type = SystemErrorCodes::InvalidDateValue
+            }
+            ParseErrorKind::TooLong => {
+                error_type = SystemErrorCodes::InvalidDateValue
+            }
+            ParseErrorKind::BadFormat => {
+                error_type = SystemErrorCodes::BadDateFormat
+            }
+            ParseErrorKind::__Nonexhaustive => {
+                error_type = SystemErrorCodes::BadDateFormat
+            }
+        }
+
+        Self {
+            error_type,
+            error_content
         }
     }
 }